chore: move repository interface to domain, rename request pkg to dto
