+++ server/package.json
@@ -26,6 +26,7 @@
     "@nestjs/typeorm": "^6.1.1",
     "class-validator": "^0.9.1",
     "nestjs-config": "^1.4.0",
+    "nestjs-typeorm-paginate": "^0.1.7",
     "reflect-metadata": "^0.1.12",
     "rimraf": "^2.6.2",
     "rxjs": "^6.3.3",
@@ -40,6 +41,7 @@
     "@types/supertest": "^2.0.7",
     "husky": "^2.3.0",
     "jest": "^23.6.0",
+    "jest-plugin-context": "^2.9.0",
     "nodemon": "^1.18.9",
     "prettier": "^1.15.3",
     "supertest": "^3.4.1",
