+++ server/src/app/transaction/dto/create-transaction.dto.spec.ts
@@ -1,0 +1,7 @@
+import {CreateTransactionDto} from './create-transaction.dto';
+
+describe('CreateTransactionDto', () => {
+    it('should be defined', () => {
+        expect(new CreateTransactionDto()).toBeDefined();
+    });
+});
