--- server/src/app/transaction/request/create-transaction.request.spec.ts
@@ -1,7 +1,0 @@
-import {CreateTransactionRequest} from './create-transaction.request';
-
-describe('CreateTransactionRequest', () => {
-    it('should be defined', () => {
-        expect(new CreateTransactionRequest()).toBeDefined();
-    });
-});
