+++ server/src/app/user/dto/signUp.dto.spec.ts
@@ -1,0 +1,7 @@
+import { SignUpDto } from './signUp.dto';
+
+describe('SignUpDto', () => {
+  it('should be defined', () => {
+    expect(new SignUpDto({address: 'address', message: 'message'}, 'signature')).toBeDefined();
+  });
+});
