+++ server/src/app/user/dto/signUp.dto.ts
@@ -1,5 +1,4 @@
-export class SignUpPayload {
-
+export class SignUpDto {
     data: {
         address: string,
         message: string,
