--- server/src/app/user/dto/signUp.payload.spec.ts
@@ -1,7 +1,0 @@
-import { SignUpPayload } from './signUp.payload';
-
-describe('SignUpPayload', () => {
-  it('should be defined', () => {
-    expect(new SignUpPayload({address: 'address', message: 'message'}, 'signature')).toBeDefined();
-  });
-});
