+++ server/src/app/user/user.service.impl.ts
@@ -3,16 +3,16 @@ import {UserService} from './user.service';
 import {UserDto} from '../../domain/user/dto/user.dto';
 import {User} from '../../domain/user/user.entity';
 import {DeleteResult} from 'typeorm';
-import {IUserRepository} from '../../port/persistence/repository/user.repository';
+import {IUserRepository} from '../../domain/user/user.repository';
 
 @Injectable()
 export class UserServiceImpl implements UserService {
-
     constructor(@Inject('IUserRepository') private userRepository: IUserRepository) {}
 
     async get(id: number): Promise<User> {
         return await this.userRepository.findById(id);
     }
+    
     async create(userDto: UserDto): Promise<User> {
         if (userDto.address === undefined) {
             throw new Error('address should be defined');
@@ -28,9 +28,11 @@ export class UserServiceImpl implements UserService {
         const user = new User(userDto.address, userDto.name);
         return await this.userRepository.save(user);
     }
+    
     async delete(id: number): Promise<DeleteResult> {
         return await this.userRepository.delete(id);
     }
+    
     async increasePoint(id: number, amount: number): Promise<User> {
         let user: User;
         user = await this.userRepository.findById(id);
@@ -40,6 +42,7 @@ export class UserServiceImpl implements UserService {
 
         return user.increasePoint(amount);
     }
+    
     async decreasePoint(id: number, amount: number): Promise<User> {
         let user: User;
         user = await this.userRepository.findById(id);
