+++ server/src/domain/transaction/transaction.repository.ts
@@ -1,5 +1,5 @@
-import {Transaction} from '../../../domain/transaction/transaction.entity';
 import {FindManyOptions} from 'typeorm';
+import {Transaction} from './transaction.entity';
 
 export interface ITransactionRepository {
     find(options?: FindManyOptions<Transaction>): Promise<Transaction[]>;
