+++ server/src/port/adapter/service/blockchain/ethereum/web3/web3.bridge.service.ts
@@ -1,7 +1,7 @@
 import {BridgeService} from '../../../../../../domain/user/bridge.service';
 import {Inject, Injectable} from '@nestjs/common';
 import Web3 from 'web3';
-import {SignUpPayload} from '../../../../../../app/user/dto/signUp.payload';
+import {SignUpDto} from '../../../../../../app/user/dto/signUp.dto';
 
 @Injectable()
 export class Web3BridgeService implements BridgeService {
