+++ server/src/port/persistence/repository/transaction.repository.impl.ts
@@ -1,6 +1,6 @@
 import {EntityRepository, Repository} from 'typeorm';
 import {Transaction} from '../../../domain/transaction/transaction.entity';
-import {ITransactionRepository} from './transaction.repository';
+import {ITransactionRepository} from '../../../domain/transaction/transaction.repository';
 
 @EntityRepository(Transaction)
 export class TransactionRepository extends Repository<Transaction> implements ITransactionRepository {}
