+++ server/src/port/persistence/repository/user.repository.impl.ts
@@ -1,10 +1,9 @@
 import {EntityRepository, Repository} from 'typeorm';
 import {User} from '../../../domain/user/user.entity';
-import {IUserRepository} from './user.repository';
+import {IUserRepository} from '../../../domain/user/user.repository';
 
 @EntityRepository(User)
 export class UserRepository extends Repository<User> implements IUserRepository {
-
     async findById(id: number): Promise<User> {
         return await this.findOne(id);
     }
