+++ server/src/web/transaction/transaction.contoller.ts
@@ -1,7 +1,7 @@
-import {Body, Controller, Get, Inject, Injectable, Param, Post, Query, Req} from '@nestjs/common';
+import {Body, Controller, Get, Inject, Injectable, Param, Post, Query} from '@nestjs/common';
 import {TransactionService} from '../../app/transaction/transaction.service.impl';
 import {Transaction} from '../../domain/transaction/transaction.entity';
-import {CreateTransactionRequest} from '../../app/transaction/request/create-transaction.request';
+import {CreateTransactionDto} from '../../app/transaction/dto/create-transaction.dto';
 
 @Injectable()
 @Controller('transactions')
@@ -19,7 +19,7 @@ export class TransactionController {
     }
 
     @Post()
-    async create(@Body() createTransactionRequest: CreateTransactionRequest): Promise<Transaction> {
-        return await this.service.create(createTransactionRequest);
+    async create(@Body() createTransactionDto: CreateTransactionDto): Promise<Transaction> {
+        return await this.service.create(createTransactionDto);
     }
 }
